# api/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.loginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: 博客 API。登录使用同源 HTTPS HttpOnly Cookie，前台和后台会话独立。先调用对应登录接口，再使用 Try it
    out；不使用 Bearer Token。当前文档覆盖健康检查、登录会话、文章及写作工作流。
  title: Kakozane Blog API
  version: "1.0"
paths:
  /admin/auth/login:
    post:
      consumes:
      - application/json
      description: Cookie 由浏览器自动管理；写操作必须通过同源 HTTPS 发起。
      parameters:
      - description: 账号密码
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: user 信息；仅后台包含 role、permissions
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "426":
          description: Upgrade Required
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 后台认证 - Login
      tags:
      - 后台认证
  /admin/auth/logout:
    post:
      description: Cookie 由浏览器自动管理；写操作必须通过同源 HTTPS 发起。
      produces:
      - application/json
      responses:
        "204":
          description: 已退出
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "426":
          description: Upgrade Required
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 后台认证 - Logout
      tags:
      - 后台认证
  /admin/auth/me:
    get:
      description: Cookie 由浏览器自动管理；写操作必须通过同源 HTTPS 发起。
      produces:
      - application/json
      responses:
        "200":
          description: user 信息；仅后台包含 role、permissions
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "426":
          description: Upgrade Required
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 后台认证 - Me
      tags:
      - 后台认证
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Cookie 由浏览器自动管理；写操作必须通过同源 HTTPS 发起。
      parameters:
      - description: 账号密码
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: user 信息；仅后台包含 role、permissions
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "426":
          description: Upgrade Required
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 前台认证 - Login
      tags:
      - 前台认证
  /auth/logout:
    post:
      description: Cookie 由浏览器自动管理；写操作必须通过同源 HTTPS 发起。
      produces:
      - application/json
      responses:
        "204":
          description: 已退出
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "426":
          description: Upgrade Required
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 前台认证 - Logout
      tags:
      - 前台认证
  /auth/me:
    get:
      description: Cookie 由浏览器自动管理；写操作必须通过同源 HTTPS 发起。
      produces:
      - application/json
      responses:
        "200":
          description: user 信息；仅后台包含 role、permissions
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "426":
          description: Upgrade Required
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too Many Requests
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 前台认证 - Me
      tags:
      - 前台认证
...
schemes:
- https
swagger: "2.0"